requirements.txt
# Flask
flask==3.0.0
Flask-Cors==4.0.0

# Web
Werkzeug==3.0.1

# data analysis
XlsxWriter==3.1.9
openpyxl==3.1.2
pandas==2.1.3
# polars==0.19.19

# docs
flasgger==0.9.7.1
marshmallow==3.20.1
apispec==6.3.0

# tests
coverage==7.3.1
requests==2.31.0
locust==2.16.1

# Utils
python-dotenv==1.0.0
validate-dns-email==0.1.0